+++ uikit/src/main/java/com/sendbird/uikit/consts/DialogEditTextParams.kt
@@ -1,0 +1,13 @@
+package com.sendbird.uikit.consts
+
+import android.text.TextUtils.TruncateAt
+
+open class DialogEditTextParams private constructor(
+    var hintText: String? = null,
+    var enableSingleLine: Boolean = false,
+    var ellipsis: TruncateAt? = null,
+    var selection: Int = 0,
+    var text: String? = null,
+) {
+    constructor(hintText: String) : this(hintText = hintText, false, null, 0, null)
+}