+++ uikit/src/main/java/com/sendbird/uikit/model/MentionConfig.kt
@@ -1,0 +1,8 @@
+package com.sendbird.uikit.model
+
+/**
+ * Describes a configuration of mention
+ *
+ * @since 3.0.0
+ */
+abstract class MentionConfig