+++ uikit/src/main/java/com/sendbird/uikit/model/VoiceMessageInfo.kt
@@ -1,0 +1,7 @@
+package com.sendbird.uikit.model
+
+data class VoiceMessageInfo(
+    val path: String,
+    val mimeType: String,
+    val duration: Int
+)